requirements.txt
aiodocker==0.21.0
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
annotated-types==0.7.0
astunparse==1.6.3
atomicwrites==1.4.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
check-jsonschema==0.22.0
click==8.1.7
click-default-group==1.2.4
click-didyoumean==0.3.0
croniter==2.0.1
databricks-sql-connector==2.9.4
distlib==0.3.8
dlt==0.4.6
duckdb==0.9.2
et-xmlfile==1.1.0
fasteners==0.19
filelock==3.13.1
flatten-dict==0.4.2
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
giturlparse==0.12.0
hexbytes==1.1.0
-e git+ssh://git@github.com/Bartman0/hubspot-api-python.git@14c7516bed867fca93afa6fb1bc8266ec92bda05#egg=hubspot_api_client
humanize==4.9.0
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.3
jsonpath-ng==1.6.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.3.0
lz4==4.3.3
makefun==1.15.2
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
orjson==3.9.10
packaging==23.2
pandas==2.2.1
pathvalidate==3.2.0
pendulum==3.0.0
pillow==10.2.0
platformdirs==4.2.0
ply==3.11
prompt-toolkit==3.0.36
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.1
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.4
python-ulid==1.1.0
pytz==2024.1
PyYAML==6.0.1
questionary==2.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==2.0.0
requirements-parser==0.5.0
rich==13.7.0
rpds-py==0.17.1
ruamel.yaml==0.17.21
semver==3.0.2
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
snowplow-tracker==1.0.1
SQLAlchemy==1.4.52
streamlit==1.32.0
structlog==24.1.0
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
thrift==0.16.0
time-machine==2.14.0
toml==0.10.2
tomlkit==0.12.4
toolz==0.12.1
tornado==6.4
tqdm==4.62.3
twinfield==2.1.2
types-setuptools==69.1.0.20240308
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
virtualenv==20.25.0
wcwidth==0.2.13
yarl==1.9.4
yaspin==2.5.0
